feat: added the transaction status on the transactions table
